Typing.
