+++ pyrene/src/components/SimpleTable/SimpleTableActionList.tsx
@@ -1,10 +1,23 @@
-import React, { useState } from 'react';
-import PropTypes from 'prop-types';
+/* eslint-disable react/prop-types */
+import React, { useState, FunctionComponent } from 'react';
 import Popover from '../Popover/Popover';
 import styles from './simpleTableActionList.css';
 import Icon from '../Icon/Icon';
+import { Action } from './types';
 
-const SimpleTableActionList = (props) => {
+interface SimpleTableActionListProps {
+  actions: Array<Action>,
+  row: {
+    key?: string,
+    rowStyle?: any,
+    value?: React.ReactNode,
+  },
+}
+
+const SimpleTableActionList: FunctionComponent<SimpleTableActionListProps> = ({
+  actions,
+  row,
+}) => {
 
   const [activeAction, setActiveAction] = useState({
     displayed: false,
@@ -18,7 +31,7 @@ const SimpleTableActionList = (props) => {
       distanceToTarget={14}
       renderPopoverContent={() => (
         <div className={styles.actionMenuContainer}>
-          {props.actions.map((element) => (
+          {actions.map((element) => (
             <a
               className={styles.actionLink}
               key={`${element.label}_actionLink`}
@@ -27,7 +40,7 @@ const SimpleTableActionList = (props) => {
                 if (element.onClick) {
                   e.preventDefault();
                   e.stopPropagation();
-                  element.onClick(props.row);
+                  element.onClick(row);
                   onClose();
                 }
               }}
@@ -56,18 +69,6 @@ const SimpleTableActionList = (props) => {
       </div>
     </Popover>
   );
-};
-
-SimpleTableActionList.propTypes = {
-  actions: PropTypes.arrayOf(PropTypes.shape({
-    label: PropTypes.string.isRequired,
-    onClick: PropTypes.func.isRequired,
-  })).isRequired,
-  row: PropTypes.shape({
-    key: PropTypes.string,
-    rowStyle: PropTypes.shape({}),
-    value: PropTypes.node,
-  }).isRequired,
 };
 
 export default SimpleTableActionList;
